requirements.txt
bert==2.2.0
boto3==1.26.70
botocore==1.29.70
h5py==3.2.1
kfac==0.2.4
nltk==3.8.1
numpy==1.23.5
pandas==1.5.3
ptvsd==4.3.2
pytorch_pretrained_bert==0.6.2
Requests==2.28.2
scikit_learn==1.2.2
scipy==1.10.1
six==1.16.0
tensorflow==2.12.0
tokenizers==0.10.1
tqdm==4.65.0
wget==3.2
